+++ src/@types/project.d.ts
@@ -1,8 +1,8 @@
 type ProjectInfo = {
   name        : string,
   status?     : string,  // open or close
-  opened?     : string,
-  closed?     : string,
+  opened?     : string,  // opened date
+  closed?     : string,  // closed date
   description : string,
   index       : number,
   bundles     : Array<BundleInfo>
@@ -12,6 +12,7 @@ type BundleInfo = {
   id          : number,
   name        : string,
   description : string,
+  date?       : string,  // collected date
   available   : boolean
 }
 
@@ -22,5 +23,5 @@ type FileInfo = {
   isDirectory : boolean,
   children    : Array<string>,
   size        : number,
-  modifiedAt  : Date
+  mtime       : string
 }