+++ src/client/components/complexes/bundle-upload-modal.tsx
@@ -78,14 +78,21 @@ const BundleUploadBox = React.memo<BundleUploadBoxProps>(({
       status.current.success = true
       status.current.progress = 0
       clearFrom()
+      return
     })
-    .catch((err: AxiosError) => {
-      data.current.message = err.response.data.msg
+    .catch((err: Error | AxiosError) => {
+      if (Axios.isAxiosError(err)) {
+        data.current.message = err.response.data.msg
+      } else {
+        data.current.message = "An error on the client occurred."
+        console.log(err)
+      }
       status.current.processing = false
       status.current.done = true
       status.current.success = false
       status.current.progress = 0
       clearFrom()
+      return
     })
   }, [domain, project])