+++ src/client/components/complexes/markdown-viewer-box.tsx
@@ -36,8 +36,12 @@ const MarkdownViewerBox = React.memo<MarkdownViewerBoxProps>(({
       forceUpdate()
       return
     })
-    .catch((err: AxiosError) => {
-      alert(err.response.data.msg)
+    .catch((err: Error | AxiosError) => {
+      if (Axios.isAxiosError(err)) {
+        alert(err.response.data.msg)
+      } else {
+        console.log(err)
+      }
       return
     })
   }, [true])